backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.4
django-appconf==1.0.6
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
kombu==5.3.7
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
